Reworked precision telemetry value parsing logic
...

+++ src/serial/Messages/Message.java
@@ -99,7 +99,7 @@ public class Message {
         return new WordMessage(Serial.SYNC_WORD, resync, (byte)0);
     }
     
-    public static Message telemetry(byte index, float data) {
+    public static Message telemetry(byte index, double data) {
         return new DataMessage(Serial.TELEMETRY_DATA, index, data);
     }
     
@@ -107,7 +107,7 @@ public class Message {
         return new DataMessage(Serial.SETTING_DATA, index, data);
     }
     
-    public static Message sensor(byte index, float data) {
+    public static Message sensor(byte index, double data) {
     	return new DataMessage(Serial.SENSOR_DATA, index, data);
     }
